Gracefully handle errors in usercode
Until now if there was an error in the usercode we wouldn't track that
the code didn't completely execute, then the next code diff would just
move on as if everything was peachy. This is wrong. This patch takes a
database transaction approach where it will only commit the new ast if
all the expressions needed to run completed successfully.

You can conceive a few examples where this isn't what you want, but if
the code changes between saves are small enough, which I believe is the
behavior we shuold encourage, then this right decision. But! if i'm
wrong we can always change it.

This patch also does some slight internal refactoring to make it easier
to test the actual control loop.